+++ src/index.js
@@ -1,7 +1,7 @@
 // @flow
 
 import { createConnection, type Socket } from 'net'
-import { EOL } from 'os'
+import oboe from 'oboe'
 import {
   type Observer,
   ReplaySubject,
@@ -14,41 +14,29 @@ import { AnonymousSubject } from 'rxjs/internal/Subject'
 type OpenObserver = Observer<void>
 type CloseObserver = Observer<boolean>
 
-type ConnectArg = string | net$connectOptions
+export type ConnectArg = string | net$connectOptions
 
-export type Config<T> = {
+export type Config = {
   connect: ConnectArg,
-  deserializer: (value: string) => T,
-  serializer: (value: any) => string,
   openObserver?: ?OpenObserver,
   closeObserver?: ?CloseObserver,
-  frameSize: number,
-  maxBufferFrames: number,
 }
 
-const DEFAULT_CONFIG = {
-  deserializer: (value: string) => JSON.parse(value),
-  serializer: (value: any) => JSON.stringify(value),
-  frameSize: 8192,
-  maxBufferFrames: 10,
-}
+export type ConnectOrConfig = ConnectArg | Config
 
 export class SocketSubject<T> extends AnonymousSubject<T> {
-  _config: Config<T>
+  _config: Config
   _output: Subject<T>
   _socket: ?Socket
-  _buffer: string = ''
 
-  constructor(connectOrConfig: ConnectArg | Config<T>) {
+  constructor(connectOrConfig: ConnectOrConfig) {
     super()
 
-    const config =
+    this._config =
+      // $FlowFixMe: config type
       typeof connectOrConfig === 'string' || connectOrConfig.connect == null
         ? { connect: connectOrConfig }
         : connectOrConfig
-
-    // $FlowFixMe: config type
-    this._config = Object.assign({}, DEFAULT_CONFIG, config)
     this._output = new Subject()
     this.destination = new ReplaySubject()
   }
@@ -59,18 +47,8 @@ export class SocketSubject<T> extends AnonymousSubject<T> {
   }
 
   _connectSocket() {
-    const {
-      connect,
-      deserializer,
-      serializer,
-      openObserver,
-      closeObserver,
-      frameSize,
-      maxBufferFrames,
-    } = this._config
+    const { connect, openObserver, closeObserver } = this._config
     const observer = this._output
-
-    const maxBufferSize = frameSize * maxBufferFrames
 
     const socket = createConnection(connect)
     this._socket = socket
@@ -84,7 +62,7 @@ export class SocketSubject<T> extends AnonymousSubject<T> {
 
       this.destination = Subscriber.create(
         x => {
-          socket.write(serializer(x))
+          socket.write(JSON.stringify(x))
         },
         e => {
           observer.error(e)
@@ -111,78 +89,19 @@ export class SocketSubject<T> extends AnonymousSubject<T> {
       }
     })
 
-    const tryPush = (value: T): void => {
-      try {
-        observer.next(value)
-      } catch (err) {
-        observer.error(err)
-      }
-    }
-
-    const tryParse = (value: string): ?T => {
-      try {
-        return deserializer(value)
-      } catch (err) {
-        // Swallow error
-      }
-    }
-
-    const tryParseAll = (value: string): Array<T> => {
-      return value
-        .split(EOL)
-        .map(tryParse)
-        .filter(Boolean)
-    }
-
-    socket.on('data', data => {
-      const str = data.toString()
-      let parsed = []
-
-      if (str.length === frameSize) {
-        // Frame is full: could be incomplete message
-        if (this._buffer.length > 0) {
-          // A previous frame is already buffered, try to parse them combined
-          parsed = tryParseAll(this._buffer + str)
-          if (parsed.length === 0) {
-            // No message parsed with combined frames, try with only the new frame
-            parsed = tryParseAll(str)
-            if (parsed.length === 0) {
-              // Failed to parse a message from new frame
-              if (this._buffer.length === maxBufferSize) {
-                // Buffer is full, emit error
-                observer.error(new Error('Buffer overflow'))
-              } else {
-                // Append frame to existing buffer
-                this._buffer += str
-              }
-            }
-          }
-        } else {
-          // No previous frame in buffer, try to parse new one
-          parsed = tryParseAll(str)
-          if (parsed.length === 0) {
-            // Failed to parse a message from new frame, buffer it
-            this._buffer = str
-          }
+    oboe(socket)
+      .on('done', (value: T) => {
+        try {
+          observer.next(value)
+        } catch (err) {
+          observer.error(err)
         }
-      } else if (this._buffer.length > 0) {
-        // A previous frame is already buffered, try to parse them combined
-        parsed = tryParseAll(this._buffer + str)
-        if (parsed.length === 0) {
-          // No message parsed with combined frames, try with only the new frame
-          parsed = tryParseAll(str)
-        }
-      } else {
-        // Basic case: frame is not full and buffer is empty
-        parsed = tryParseAll(str)
-      }
-
-      if (parsed.length !== 0) {
-        // At least one message was parsed, clear buffer and push messages
-        this._buffer = ''
-        parsed.forEach(tryPush)
-      }
-    })
+      })
+      .on('fail', (report: { thrown: ?Error, body: ?string }) => {
+        observer.error(
+          report.thrown || new Error(report.body || 'Socket error'),
+        )
+      })
   }
 
   _subscribe(subscriber: Subscriber<T>): Subscription {